db_monitor/requirements.txt
requests
boto3
#base64
#os
mysql-connector-python